fix: avoid exposing internal flags on $scopedSlots
ref #9443
